First iteration of rectify_fpga
Validated in the FPGA the remap function resulting in
consistent responses when comparing PS and PL operations.

Next up is to consider the initialization of the
rectification maps and either keep that in the PS or push
it to the PL as well.

Adjustments in the CMakeLists.txt and a new Node for
testing purposes have been added.

Signed-off-by: Víctor Mayoral Vilches <v.mayoralv@gmail.com>