Second iteration of rectify_fpga
Restructure code and subclass PinholeCameraModel to reuse
that package's logic. This commit shows how both the kernel
and the OpenCV functions provide the same result when
facilitated the same warp map.

Only remap is pushed to the acceleration kernel because
cv::initRectificationMaps is negligible in comparison. In
the future, offloading this function also to the FPGA can
be considered by using xf::cv::InitUndistortRectifyMapInverse.

Signed-off-by: Víctor Mayoral Vilches <v.mayoralv@gmail.com>